PetDx/backend/requirements.txt
flask
tensorflow
numpy
pillow
flask-cors
pymongo
python-dotenv
cloudinary